pyproject.toml
[project]
name = "1-anime-recommender-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.2.1",
    "langchain>=1.0.2",
    "langchain-community>=0.4",
    "langchain-groq>=1.0.0",
    "langchain-huggingface>=1.0.0",
    "langchain-text-splitters>=1.0.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.50.0",
]
